Update backend.ps1
I modified the code slightly to take advantage of the common functions, and to remove a redundant stop I found on Friday.  I have not tested this code, so please do so as soon as possible and check in any changes.

Please see the two TODOs that I put in and, if possible, implement them.

Does this create a new back end factory every time the thing is called?  It looked like it did.

If the back end factory is persistent at least for the duration of a script, and it kept track of all the individual machines it has launched, it could solve the problem I'm having of creating and tearing down PSRP (PowerShell remoting) sessions all the time.  If I could do something like $session = ($backend.findInstance($vmName)).GetPSSession(), that would be cool!

We need a companion to launch_single_azure_vm that is destroy_single_azure_vm.  That would be called by stop_machines_in_group and/or deallocate_machines_in_group from common_functions.